requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
mysql-connector-python
dotenv
pytest
PyJWT